+++ src/plugins/stake-plugins/stakePlugins.ts
@@ -1,5 +1,5 @@
 import ENV from '../../../env.json'
-import { makeTcSaversPlugin } from './thorchainSavers/tcSavers'
+import { makeTcSaversPlugin } from './thorchainSavers/tcSaversPlugin'
 import { StakePlugin } from './types'
 import { makeUniV2StakePlugin } from './uniswapV2/uniV2Plugin'
